+++ packages/sdk/src/NocturneWalletSDK.ts
@@ -6,25 +6,23 @@ import {
 import { NocturneSigner } from "./crypto";
 import { handleGasForOperationRequest } from "./opRequestGas";
 import { MerkleProver } from "./merkleProver";
-import { OpPreparer } from "./opPreparer";
+import { prepareOperation } from "./prepareOperation";
 import { OperationRequest } from "./operationRequest";
 import { NocturneDB } from "./NocturneDB";
 import { Wallet, Wallet__factory } from "@nocturne-xyz/contracts";
 import { ethers } from "ethers";
-import { OpSigner } from "./opSigner";
+import { signOperation } from "./signOperation";
 import { loadNocturneConfig, NocturneConfig } from "@nocturne-xyz/config";
 import { Asset, AssetTrait } from "./primitives/asset";
 import { SyncAdapter } from "./sync";
-import { NocturneSyncer } from "./NocturneSyncer";
+import { syncSDK } from "./syncSDK";
+import { getJoinSplitRequestTotalValue } from "./utils";
 
-export class NocturneContext {
-  private opPreparer: OpPreparer;
-  private opSigner: OpSigner;
-  private syncer: NocturneSyncer;
-
+export class NocturneWalletSDK {
   protected walletContract: Wallet;
   protected merkleProver: MerkleProver;
   protected db: NocturneDB;
+  protected syncAdapter: SyncAdapter;
 
   readonly signer: NocturneSigner;
   readonly gasAssets: Asset[];
@@ -55,40 +53,37 @@ export class NocturneContext {
     );
     this.merkleProver = merkleProver;
     this.db = db;
-    this.opPreparer = new OpPreparer(this.db, this.merkleProver, this.signer);
-
-    this.opSigner = new OpSigner(this.signer);
-
-    this.syncer = new NocturneSyncer(
-      this.signer,
-      syncAdapter,
-      this.db,
-      this.merkleProver,
-      provider
-    );
+    this.syncAdapter = syncAdapter;
   }
 
   async sync(): Promise<void> {
-    await this.syncer.sync();
+    const deps = {
+      provider: this.walletContract.provider,
+      viewer: this.signer,
+    };
+    await syncSDK(deps, this.syncAdapter, this.db, this.merkleProver);
   }
 
   async prepareOperation(
     opRequest: OperationRequest
   ): Promise<PreSignOperation> {
+    const deps = {
+      db: this.db,
+      gasAssets: this.gasAssets,
+      walletContract: this.walletContract,
+      merkle: this.merkleProver,
+      viewer: this.signer,
+    };
+
     const gasAccountedOpRequest = await handleGasForOperationRequest(
-      {
-        db: this.db,
-        gasAssets: this.gasAssets,
-        opPreparer: this.opPreparer,
-        walletContract: this.walletContract,
-      },
+      deps,
       opRequest
     );
-    return await this.opPreparer.prepareOperation(gasAccountedOpRequest);
+    return await prepareOperation(deps, gasAccountedOpRequest);
   }
 
   signOperation(preSignOperation: PreSignOperation): SignedOperation {
-    return this.opSigner.signOperation(preSignOperation);
+    return signOperation(this.signer, preSignOperation);
   }
 
   async getAllAssetBalances(): Promise<AssetWithBalance[]> {
@@ -106,6 +101,16 @@ export class NocturneContext {
   async hasEnoughBalanceForOperationRequest(
     opRequest: OperationRequest
   ): Promise<boolean> {
-    return this.opPreparer.hasEnoughBalanceForOperationRequest(opRequest);
+    for (const joinSplitRequest of opRequest.joinSplitRequests) {
+      const requestedAmount = getJoinSplitRequestTotalValue(joinSplitRequest);
+      // check that the user has enough notes to cover the request
+      const notes = await this.db.getNotesForAsset(joinSplitRequest.asset);
+      const balance = notes.reduce((acc, note) => acc + note.value, 0n);
+      if (balance < requestedAmount) {
+        return false;
+      }
+    }
+
+    return true;
   }
 }