+++ packages/snap/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- use new `NocturneWalletSDK` instead of `NocturneContext`
 - make `SnapKVStore` only dump on writes by putting the inner KV store in a thunk
 - use `SubgraphSyncAdapter` against local graph node
 - call `context.sync()` at the beginning of `nocturne_signOperation` so it's unnecessary for frontend to explicitly sync